requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
dash==0.20.0
dash-core-components==0.18.1
dash-html-components==0.8.0
dash-renderer==0.11.2
decorator==4.0.11
Flask==0.12.2
Flask-Compress==1.4.0
idna==2.6
imageio==2.1.2
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.0
moviepy==0.2.3.2
nbformat==4.4.0
numpy==1.14.0
oauthlib==2.0.6
pandas==0.22.0
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
plotly==2.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
schedule==0.5.0
simplegeneric==0.8.1
six==1.11.0
tqdm==4.11.2
traitlets==4.3.2
tweepy==3.5.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
